lunaris-orion/models/requirements.txt
torch==2.0.0
torchvision==0.15.0
Pillow==10.2.0
tqdm==4.66.1
numpy==1.24.3
tensorboard==2.15.1
transformers==4.36.2
wandb==0.16.2
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.4
opencv-python==4.9.0.80
albumentations==1.3.1
requests>=2.31.0 